[PR] Made FGDemoView into a ViewController
